+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 [Upgrade Instructions](https://maru.readme.io/v0.12/docs/upgrade-instructions-from-v011) From v0.11 to v0.12
 
 ## v0.12.4-dev
+* Enhancements
+  * make regexp validator support list
 * Bugfix
   * Support :chunk response